backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.10.0
httpx==0.26.0
websockets==12.0
supabase==2.3.0
pandas==2.1.4
numpy==1.26.0
python-binance==1.0.19
ta-lib==0.4.28
schedule==1.2.0
aiocron==1.8
